+++ src/api/BoreholeLockService.cs
@@ -5,7 +5,7 @@ namespace BDMS;
 /// <inheritdoc />
 public class BoreholeLockService(BdmsContext context, ILogger<BoreholeLockService> logger, TimeProvider timeProvider) : IBoreholeLockService
 {
-    internal const int LockTimeoutInMinutes = 10;
+    internal const int LockTimeoutInMinutes = 60;
 
     private readonly BdmsContext context = context;
     private readonly ILogger<BoreholeLockService> logger = logger;
@@ -19,6 +19,9 @@ public class BoreholeLockService(BdmsContext context, ILogger<BoreholeLockServic
             .AsNoTracking()
             .SingleOrDefaultAsync(u => u.SubjectId == subjectId)
             .ConfigureAwait(false) ?? throw new InvalidOperationException($"Current user with subjectId <{subjectId}> does not exist.");
+
+        // Admins can always edit boreholes
+        if (user.IsAdmin) return false;
 
         var borehole = await context.Boreholes
             .Include(b => b.Workflows)